+++ NEWS.md
@@ -1,4 +1,7 @@
 # wright NEWS
+## unreleased
+- Add --dry-run option
+
 ## 0.3.0 (2015-04-23)
 - Add bin/wright
 - Add wright(1) manpage